refactor: factor out ClassificationHandler from PotreeNodeWrapper
