+++ viewer/packages/pointclouds/src/PotreeNodeWrapper.ts
@@ -4,20 +4,16 @@
 
 import * as THREE from 'three';
 
-const PotreeDefaultPointClass = 'DEFAULT';
-
-type ClassificationMap = { [key: string]: { rgb: THREE.Color; code: number } };
-
 import { StyledPointCloudObjectCollection, CompletePointCloudAppearance } from '@reveal/pointcloud-styling';
 
 import { PointCloudOctree, PickPoint } from './potree-three-loader';
-import { PointColorType, PointShape, PointClassification, DEFAULT_CLASSIFICATION } from '@reveal/rendering';
+import { PointColorType, PointShape, PointClassification } from '@reveal/rendering';
 
 import { WellKnownAsprsPointClassCodes } from './types';
 
 import { PointCloudObjectMetadata, PointCloudObject } from '@reveal/data-providers';
 import { ClassificationInfo } from './potree-three-loader/loading/ClassificationInfo';
-import { createDistinctColors } from '@reveal/utilities';
+import { ClassificationHandler } from './ClassificationHandler';
 
 /**
  * Wrapper around `Potree.PointCloudOctree` with some convenience functions.
@@ -25,14 +21,14 @@ import { createDistinctColors } from '@reveal/utilities';
 export class PotreeNodeWrapper {
   readonly octree: PointCloudOctree;
   private _needsRedraw = false;
-  private readonly _classification: PointClassification;
 
   private readonly _modelIdentifier: symbol;
 
   private static readonly pickingWindowSize = 20;
-  private readonly _classNameToCodeMap: ClassificationMap;
 
   private readonly _annotations: PointCloudObject[];
+
+  private readonly _classificationHandler: ClassificationHandler;
 
   get needsRedraw(): boolean {
     return this._needsRedraw;
@@ -48,64 +44,10 @@ export class PotreeNodeWrapper {
     this.pointSize = 2;
     this.pointColorType = PointColorType.Rgb;
     this.pointShape = PointShape.Circle;
-    this._classification = octree.material.classification;
     this._annotations = annotations;
     this._modelIdentifier = modelIdentifier;
-    this._classNameToCodeMap = this.createClassNameToCodeMap(classificationInfo);
-
-    this.updateMaterialClassMap(this._classNameToCodeMap);
-  }
-
-  private createClassNameToCodeMap(classificationInfo: ClassificationInfo): ClassificationMap {
-    if (classificationInfo.classificationSets.length === 0) {
-      return this.createDefaultClassNameToCodeMap();
-    }
-
-    const classMap: { [key: string]: { rgb: THREE.Color; code: number } } = {};
-
-    const inputClassifications = classificationInfo.classificationSets[0].classificationSet;
-
-    const fallbackColors = createDistinctColors(inputClassifications.length);
-
-    inputClassifications.forEach((pointClass, index) => {
-      const rgb = pointClass.rgb ? new THREE.Color(pointClass.rgb) : fallbackColors[index];
-
-      classMap[pointClass.name] = { rgb, code: pointClass.code };
-    });
-
-    return classMap;
-  }
-
-  private getDefaultClassColor(name: string): THREE.Color | undefined {
-    if (!DEFAULT_CLASSIFICATION[name]) {
-      return undefined;
-    }
-
-    const v = DEFAULT_CLASSIFICATION[name];
-    return new THREE.Color(v.x, v.y, v.z);
-  }
-
-  private createDefaultClassNameToCodeMap(): ClassificationMap {
-    const classNames = Object.keys(this._classification);
-
-    const classMap: { [key: string]: { rgb: THREE.Color; code: number } } = {};
-    const fallbackColors = createDistinctColors(classNames.length);
-
-    classNames.forEach((x, index) => {
-      const code = x === PotreeDefaultPointClass ? -1 : parseInt(x, 10);
-      const rgb = this.getDefaultClassColor(x) ?? fallbackColors[index];
-      classMap[this.getClassNameFromCode(code)] = { code, rgb };
-    });
-
-    return classMap;
-  }
-
-  private updateMaterialClassMap(classificationMap: ClassificationMap) {
-    Object.keys(classificationMap).forEach(name => {
-      const color = classificationMap[name].rgb;
-      this._classification[classificationMap[name].code] = new THREE.Vector4(...color.toArray(), 1.0);
-    });
-    this.octree.material.classification = this._classification;
+    this._classificationHandler = new ClassificationHandler(octree.material,
+                                                            classificationInfo);
   }
 
   get modelIdentifier(): symbol {
@@ -134,6 +76,7 @@ export class PotreeNodeWrapper {
   get pointColorType(): PointColorType {
     return this.octree.material.pointColorType;
   }
+
   set pointColorType(type: PointColorType) {
     this.octree.material.pointColorType = type;
     this._needsRedraw = true;
@@ -148,20 +91,11 @@ export class PotreeNodeWrapper {
   }
 
   get classification(): PointClassification {
-    return this._classification;
-  }
-
-  private getClassNameFromCode(code: number): string {
-    return WellKnownAsprsPointClassCodes[code] ?? `Class ${code}`;
+    return this._classificationHandler.classification;
   }
 
   get classes(): Array<{ name: string; code: number }> {
-    const codesAndNames = Object.entries(this._classNameToCodeMap).map(nameAndCode => ({
-      name: nameAndCode[0],
-      code: nameAndCode[1].code
-    }));
-
-    return codesAndNames.sort((a, b) => a.code - b.code);
+    return this._classificationHandler.classes;
   }
 
   get stylableObjectAnnotationMetadata(): Iterable<PointCloudObjectMetadata> {
@@ -197,28 +131,11 @@ export class PotreeNodeWrapper {
   }
 
   createPointClassKey(pointClass: number | WellKnownAsprsPointClassCodes): number {
-    if (this._classNameToCodeMap[this.getClassNameFromCode(pointClass)] !== undefined) {
-      return this._classNameToCodeMap[this.getClassNameFromCode(pointClass)].code;
-    }
-
-    if (typeof pointClass === 'string') {
-      throw Error(`Unrecognized string ${pointClass} used as class name`);
-    }
-
-    if (pointClass === WellKnownAsprsPointClassCodes.Default) {
-      // Potree has a special class 'DEFAULT'. Our map has number keys, but this one is specially
-      // handled in Potree so we ignore type.
-      return PotreeDefaultPointClass as any;
-    }
-
-    return pointClass;
+    return this._classificationHandler.createPointClassKey(pointClass);
   }
 
   setClassificationAndRecompute(pointClass: number | WellKnownAsprsPointClassCodes, visible: boolean): void {
-    const key = this.createPointClassKey(pointClass);
-
-    this._classification[key].w = visible ? 1.0 : 0.0;
-    this.octree.material.classification = this._classification;
+    this._classificationHandler.setClassificationAndRecompute(pointClass, visible);
     this._needsRedraw = true;
   }
